feat: audit entities
