+++ src/main/java/shop/dreamtree/jlog/domain/BaseEntity.java
@@ -1,0 +1,29 @@
+package shop.dreamtree.jlog.domain;
+
+import java.time.LocalDateTime;
+
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseEntity {
+
+    @CreatedDate
+    private LocalDateTime createdDate;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedDate;
+
+    public LocalDateTime createdDate() {
+        return createdDate;
+    }
+
+    public LocalDateTime modifiedDate() {
+        return modifiedDate;
+    }
+}
